+++ YearInReview/Model/Aggregators/Data/GameWithTime.cs
@@ -5,6 +5,6 @@ namespace YearInReview.Model.Aggregators.Data
 	public class GameWithTime
 	{
 		public Game Game { get; set; }
-		public ulong TimePlayed { get; set; }
+		public int TimePlayed { get; set; }
 	}
 }
